requirement.txt
addict==2.4.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
astor==0.8.1
backcall==0.2.0
black==21.4b2
blis==0.7.4
block.bootstrap.pytorch==0.1.6
bootstrap.pytorch==0.0.13
brotlipy==0.7.0
catalogue==2.0.6
certifi==2021.5.30
channelnorm-cuda==0.0.0
cloudpickle==1.6.0
colorama==0.4.4
contextvars==2.4
correlation-cuda==0.0.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
dataclasses==0.7
decorator==4.4.2
filelock==3.0.10
future==0.18.2
fvcore==0.1.5.post20210515
google-pasta==0.2.0
hydra-core==1.1.0rc1
imageio==2.15.0
immutables==0.16
importlib-resources==5.1.3
instaboostfast==0.1.2
iopath==0.1.8
ipdb==0.13.9
ipython==7.12.0
Jinja2==3.0.2
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
language-tool-python==2.5.5
MarkupSafe==2.0.1
matplotlib==3.3.4
mmcv-full==1.4.8
model-index==0.1.11
munch==2.5.0
murmurhash==1.0.5
mypy-extensions==0.4.3
networkx==2.5.1
olefile==0.46
omegaconf==2.1.0rc1
opencv-python==4.5.1.48
openmim==0.1.5
ordered-set==4.0.2
pathspec==0.8.1
pathy==0.6.0
Pillow==8.4.0
plotly==5.3.1
portalocker==2.3.0
preshed==3.0.5
pretrainedmodels==0.7.4
protobuf==3.13.0
pycocotools==2.0.2
pydantic==1.8.2
pydot==1.4.2
pyhocon==0.3.58
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==19.0.2
resample2d-cuda==0.0.0
scikit-image==0.17.2
scikit-learn==0.24.2
seaborn==0.11.2
sk-video==1.1.10
skipthoughts==0.0.1
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
stanfordnlp==0.2.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminaltables==3.1.10
thinc==8.0.10
threadpoolctl==3.0.0
tifffile==2020.9.3
tokenizers==0.10.3
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
traitlets==4.3.3
transformers==4.11.3
typed-args==0.4.2
typed-ast==1.4.3
typer==0.4.0
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
yapf==0.32.0
numpy==1.19.2
nltk
h5py
Sharedarray